+++ NEWS.md
@@ -1,3 +1,10 @@
-# DSAM 1.0.0
+# DSAM 1.0.1  
 
-* Initial CRAN submission.
+* Fixed the bug in the getAUC() function.  
+* Fixed the bug in the DP.initialSample() function in extreme cases where Euclidean distance is 0.  
+* Updated the README file.  
+  
+  
+# DSAM 1.0.0  
+
+* Initial CRAN submission.